requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.7.0
asgiref==3.8.1
attrs==24.3.0
blosc2==3.0.0
Brotli==1.1.0
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
contourpy==1.3.1
cssselect2==0.7.0
cycler==0.12.1
deprecation==2.1.0
distro==1.9.0
Django==5.1.4
exa==0.5.26
exa_py==1.7.0
fastapi==0.115.6
fonttools==4.55.3
frozenlist==1.5.0
gotrue==2.11.0
grpclib==0.4.7
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==1.0.7
httpx==0.27.2
hyperframe==6.0.1
idna==3.10
iniconfig==2.0.0
jiter==0.8.2
jsonpatch==1.33
jsonpointer==3.0.0
kiwisolver==1.4.7
langchain==0.3.13
langchain-core==0.3.28
langchain-text-splitters==0.3.4
langgraph==0.2.60
langgraph-checkpoint==2.0.9
langgraph-sdk==0.1.48
langsmith==0.2.4
llvmlite==0.43.0
markdown-it-py==3.0.0
matplotlib==3.10.0
mdurl==0.1.2
modal==0.68.46
mpmath==1.3.0
msgpack==1.1.0
multidict==6.1.0
ndindex==1.9.2
networkx==3.4.2
numba==0.60.0
numexpr==2.10.2
numpy==2.0.2
openai==1.58.1
orjson==3.10.12
packaging==24.2
pandas==2.2.3
pillow==11.0.0
plotly==5.24.1
pluggy==1.5.0
postgrest==0.18.0
propcache==0.2.1
protobuf==5.29.2
py-cpuinfo==9.0.0
pycparser==2.22
pydantic==2.10.4
pydantic_core==2.27.2
pydyf==0.11.0
Pygments==2.18.0
pyparsing==3.2.0
pyphen==0.17.0
pytest==8.3.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
realtime==2.0.6
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
resend==2.5.1
rich==13.9.4
seaborn==0.13.2
shellingham==1.5.4
sigtools==4.0.1
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.36
sqlparse==0.5.3
starlette==0.41.3
storage3==0.9.0
supabase==2.10.0
supafunc==0.7.0
sympy==1.13.3
synchronicity==0.9.8
tables==3.10.1
tavily-python==0.5.0
tenacity==9.0.0
tiktoken==0.8.0
tinycss2==1.4.0
tinyhtml5==2.0.0
toml==0.10.2
tqdm==4.67.1
typer==0.15.1
types-certifi==2021.10.8.3
types-toml==0.10.8.20240310
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
watchfiles==1.0.3
weasyprint==63.1
webencodings==0.5.1
websockets==13.1
yarl==1.18.3
zopfli==0.2.3.post1
